+++ R/gg_hbar.R
@@ -273,7 +273,7 @@ gg_hbar <- function(data,
 #' @param y_var Unquoted variable to be on the y scale (i.e. character, factor, logical, numeric, date or datetime). If numeric, date or datetime, variable values are bins that are mutually exclusive and equidistant. Required input.
 #' @param col_var Unquoted categorical or numeric variable to colour the bars. Required input.
 #' @param text_var Unquoted variable to be used as a customised tooltip in combination with plotly::ggplotly(plot, tooltip = "text"). Defaults to NULL.
-#' @param position Whether bars are positioned by "dodge" or "stack". Defaults to "dodge".
+#' @param stack TRUE or FALSE of whether bars are to be position by "stack". Defaults to FALSE, which positions by "dodge".
 #' @param pal Character vector of hex codes. 
 #' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev Reverses the palette. Defaults to FALSE.
@@ -339,14 +339,15 @@ gg_hbar <- function(data,
 #'             x_var = body_mass_g, 
 #'             y_var = species, 
 #'             col_var = sex, 
-#'             position = "stack")
+#'             stack = TRUE)
 #'             
 gg_hbar_col <- function(data,
                         x_var,
                         y_var,
                         col_var,
                         text_var = NULL,
-                        position = NULL,
+                        stack = FALSE,
+                        # position = NULL,
                         pal = NULL,
                         pal_na = "#7F7F7F",
                         pal_rev = FALSE,
@@ -423,10 +424,10 @@ gg_hbar_col <- function(data,
   #warnings
   if (!is.numeric(x_var_vctr)) stop("Please use a numeric x variable for a horizontal bar plot")
   
-  if(!is.null(position)) {
-    if (!position %in% c("dodge", "stack")) stop("Please use a position of either 'dodge' or 'stack'")
-  }
-  
+  # if(!is.null(position)) {
+  #   if (!position %in% c("dodge", "stack")) stop("Please use a position of either 'dodge' or 'stack'")
+  # }
+  # 
   if (!is.null(col_method)) {
     if (!col_method %in% c("continuous", "bin", "quantile", "category")) stop("Please use a colour method of 'continuous', 'bin', 'quantile' or 'category'")
   }
@@ -550,10 +551,13 @@ gg_hbar_col <- function(data,
   pal_na_line <- scales::alpha(pal_na, alpha = alpha_line)
   
   #position
-  if (is.null(position)) {
-    position2 <- position_dodge2(preserve = "single")
-  }
-  else position2 <- position
+  # if (is.null(position)) {
+  #   position2 <- position_dodge2(preserve = "single")
+  # }
+  # else position2 <- position
+  
+  if (stack == FALSE) position <- position_dodge2(preserve = "single")
+  else if (stack == TRUE) position <- position_stack()
   
   #fundamentals
   plot <- ggplot(data) +
@@ -562,17 +566,26 @@ gg_hbar_col <- function(data,
     geom_col(aes(x = !!y_var, y = !!x_var, col = !!col_var, fill = !!col_var, text = !!text_var), 
              size = size_line, 
              width = size_width, 
-             position = position2) 
+             position = position) 
   
-  if (!is.null(position)) {
-    if (position == "stack") {
-      data_sum <- data %>%
-        dplyr::group_by(dplyr::across(!!y_var), .drop = FALSE) %>%
-        dplyr::summarise(dplyr::across(!!x_var, ~sum(.x, na.rm = TRUE))) %>%
-        dplyr::ungroup()
-      
-      x_var_vctr <- dplyr::pull(data_sum, !!x_var)
-    }
+  # if (!is.null(position)) {
+  #   if (position == "stack") {
+  #     data_sum <- data %>%
+  #       dplyr::group_by(dplyr::across(!!y_var), .drop = FALSE) %>%
+  #       dplyr::summarise(dplyr::across(!!x_var, ~sum(.x, na.rm = TRUE))) %>%
+  #       dplyr::ungroup()
+  #     
+  #     x_var_vctr <- dplyr::pull(data_sum, !!x_var)
+  #   }
+  # }
+  
+  if (stack == TRUE) {
+    data_sum <- data %>%
+      dplyr::group_by(dplyr::across(!!y_var), .drop = FALSE) %>%
+      dplyr::summarise(dplyr::across(!!x_var, ~sum(.x, na.rm = TRUE))) %>%
+      dplyr::ungroup()
+    
+    x_var_vctr <- dplyr::pull(data_sum, !!x_var)
   }
   
   #y scale 
@@ -1012,7 +1025,7 @@ gg_hbar_facet <- function(data,
 #' @param col_var Unquoted categorical or numeric variable to colour the bars. Required input.
 #' @param facet_var Unquoted categorical variable to facet the data by. Required input.
 #' @param text_var Unquoted variable to be used as a customised tooltip in combination with plotly::ggplotly(plot, tooltip = "text"). Defaults to NULL.
-#' @param position Whether bars are positioned by "dodge" or "stack". Defaults to "dodge".
+#' @param stack TRUE or FALSE of whether bars are to be position by "stack". Defaults to FALSE, which positions by "dodge".
 #' @param pal Character vector of hex codes. 
 #' @param pal_na The hex code or name of the NA colour to be used.
 #' @param pal_rev TRUE or FALSE of whether to reverse the pal.
@@ -1086,7 +1099,8 @@ gg_hbar_col_facet <- function(data,
                               col_var,
                               facet_var,
                               text_var = NULL,
-                              position = NULL,
+                              stack = FALSE,
+                              # position = NULL,
                               pal = NULL,
                               pal_na = "#7F7F7F",
                               pal_rev = FALSE,
@@ -1179,9 +1193,9 @@ gg_hbar_col_facet <- function(data,
     if (!col_method %in% c("continuous", "bin", "quantile", "category")) stop("Please use a colour method of 'continuous', 'bin', 'quantile' or 'category'")
   }
   
-  if(!is.null(position)) {
-    if (!position %in% c("dodge", "stack")) stop("Please use a position of either 'dodge' or 'stack'")
-  }
+  # if(!is.null(position)) {
+  #   if (!position %in% c("dodge", "stack")) stop("Please use a position of either 'dodge' or 'stack'")
+  # }
   
   #logical to factor
   if (is.logical(y_var_vctr)) {
@@ -1316,10 +1330,13 @@ gg_hbar_col_facet <- function(data,
   pal_na_line <- scales::alpha(pal_na, alpha = alpha_line)
   
   #position
-  if (is.null(position)) {
-    position2 <- position_dodge2(preserve = "single")
-  }
-  else position2 <- position
+  # if (is.null(position)) {
+  #   position2 <- position_dodge2(preserve = "single")
+  # }
+  # else position2 <- position
+  
+  if (stack == FALSE) position <- position_dodge2(preserve = "single")
+  else if (stack == TRUE) position <- position_stack()
   
   #fundamentals
   plot <- ggplot(data) +
@@ -1328,7 +1345,7 @@ gg_hbar_col_facet <- function(data,
     geom_col(aes(x = !!y_var, y = !!x_var, col = !!col_var, fill = !!col_var, text = !!text_var), 
              size = size_line, 
              width = size_width, 
-             position = position2) 
+             position = position) 
 
   #y scale
   if (is.character(y_var_vctr) | is.factor(y_var_vctr)){
@@ -1374,15 +1391,24 @@ gg_hbar_col_facet <- function(data,
   }
   
   #x scale
-  if (!is.null(position)) {
-    if (position == "stack") {
-      data_sum <- data %>%
-        dplyr::group_by(dplyr::across(c(!!y_var, !!facet_var)), .drop = FALSE) %>%
-        dplyr::summarise(dplyr::across(!!x_var, ~sum(.x, na.rm = TRUE))) %>%
-        dplyr::ungroup()
-      
-      x_var_vctr <- dplyr::pull(data_sum, !!x_var)
-    }
+  # if (!is.null(position)) {
+  #   if (position == "stack") {
+  #     data_sum <- data %>%
+  #       dplyr::group_by(dplyr::across(c(!!y_var, !!facet_var)), .drop = FALSE) %>%
+  #       dplyr::summarise(dplyr::across(!!x_var, ~sum(.x, na.rm = TRUE))) %>%
+  #       dplyr::ungroup()
+  #     
+  #     x_var_vctr <- dplyr::pull(data_sum, !!x_var)
+  #   }
+  # }
+  
+  if (stack == TRUE) {
+    data_sum <- data %>%
+      dplyr::group_by(dplyr::across(c(!!y_var, !!facet_var)), .drop = FALSE) %>%
+      dplyr::summarise(dplyr::across(!!x_var, ~sum(.x, na.rm = TRUE))) %>%
+      dplyr::ungroup()
+    
+    x_var_vctr <- dplyr::pull(data_sum, !!x_var)
   }
   
   x_zero_list <- sv_x_zero_adjust(x_var_vctr, x_balance = x_balance, x_zero = x_zero, x_zero_line = x_zero_line)